pyproject.toml
[tool.poetry]
name = "jeevanpath"
version = "0.1.0"
description = "This is a web backend for the jeevanpath.com"
authors = ["sukhioo7 <deepsukhi943@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
django = "5.2"
djangorestframework = "3.16.0"
gunicorn = "23.0.0"
pillow = "11.2.1"
whitenoise = "6.9.0"
django-cors-headers = "4.6.0"
django-ckeditor-5 = "0.2.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
